Collision fidelity greatly improved.  Will need to write additional code for intelligent scaling of 16bit samples depending on mix and max heights of terrain.  Multiple HFs with different scaling standards will be used for areas of drastically different average elevation...

+++ src/rendering/shaders/noise_test_v.glsl
@@ -45,7 +45,7 @@ void main(){
 
 
 	normal = normalize(cross(negz - posz, negx - posx));
-	gl_Position = p * v * ivec4(position, 1);
+	gl_Position = p * v * vec4(position, 1);
 
 	texCoordsOut = position.xz/10.0;
 }
